added selective product order and made changes in refferer

+++ api/src/main/java/com/community/api/endpoint/avisoft/controller/cart/CartEndPoint.java
@@ -6,6 +6,7 @@ import com.community.api.entity.CustomCustomer;
 import com.community.api.entity.CustomProduct;
 import com.community.api.services.CartService;
 import com.community.api.services.ProductReserveCategoryFeePostRefService;
+import com.community.api.services.ReserveCategoryService;
 import com.community.api.services.ResponseService;
 import com.community.api.services.SharedUtilityService;
 import com.community.api.services.exception.ExceptionHandlingImplement;
@@ -30,6 +31,9 @@ import javax.transaction.Transactional;
 import java.util.*;
 import java.util.stream.Collectors;
 
+import static com.community.api.services.ServiceProvider.ServiceProviderServiceImpl.getIntegerList;
+import static com.community.api.services.ServiceProvider.ServiceProviderServiceImpl.getLongList;
+
 @RestController
 @RequestMapping(value = "/cart",
         produces = {MediaType.APPLICATION_JSON_VALUE, MediaType.APPLICATION_XML_VALUE}
@@ -44,6 +48,8 @@ public class CartEndPoint extends BaseEndpoint {
     private CartService cartService;
     private ResponseService responseService;
     private SharedUtilityService sharedUtilityService;
+    private ReserveCategoryService reserveCategoryService;
+    private ProductReserveCategoryFeePostRefService reserveCategoryFeePostRefService;
 
     // Setter-based injection
     @Autowired
@@ -63,7 +69,15 @@ public class CartEndPoint extends BaseEndpoint {
     public void setResponseService(ResponseService responseService) {
         this.responseService = responseService;
     }
-
+    @Autowired
+    public void setReserveCategoryService(ReserveCategoryService reserveCategoryService){
+        this.reserveCategoryService=reserveCategoryService;
+    }
+    @Autowired
+    public void setReserveCategoryFeePostRefService(ProductReserveCategoryFeePostRefService reserveCategoryFeePostRefService)
+    {
+        this.reserveCategoryFeePostRefService=reserveCategoryFeePostRefService;
+    }
     @Autowired
     public void setOrderService(OrderService orderService) {
         this.orderService = orderService;
@@ -158,14 +172,34 @@ public class CartEndPoint extends BaseEndpoint {
             if(id==null)
                 return ResponseService.generateErrorResponse("Customer Id not specified",HttpStatus.BAD_REQUEST);
             Customer customer = customerService.readCustomerById(customerId);
-            if (customer == null) {
+            CustomCustomer customCustomer=entityManager.find(CustomCustomer.class,customerId);
+            if (customer == null||customCustomer==null) {
                 return ResponseService.generateErrorResponse("Customer not found for this Id", HttpStatus.NOT_FOUND);
+            }
+            if (customer.getFirstName() == null ||
+                    customer.getLastName() == null ||
+                    customer.getEmailAddress() == null ||
+                    customCustomer.getCategory() == null ||
+                    customer.getUsername() == null ||
+                    customer.getPassword() == null)
+            {
+                return ResponseService.generateErrorResponse(
+                        "All fields must be completed: First Name, Last Name, Primary Email, Username, Password, and Category are required before setting up the cart.",
+                        HttpStatus.BAD_REQUEST
+                );
             }
             Order cart = orderService.findCartForCustomer(customer);
             if (cart == null) {
                 cart = orderService.createNewCartForCustomer(customer);
             }
             Product product = catalogService.findProductById(productId);
+            Long reserveCategoryId=reserveCategoryService.getCategoryByName(customCustomer.getCategory()).getReserveCategoryId();
+            if(reserveCategoryService.getReserveCategoryFee(productId,reserveCategoryId)==null)
+                return ResponseService.generateErrorResponse("Cannot add product to cart :Fee not specified for your category",HttpStatus.UNPROCESSABLE_ENTITY);
+            /*if(productReserveCategoryFeePostRefService.getCustomProductReserveCategoryFeePostRefByProductIdAndReserveCategoryId(product.getId(),.getFee()==null)
+            {
+
+            }*/
             if (product == null) {
                 return ResponseService.generateErrorResponse("Product not found", HttpStatus.NOT_FOUND);
             }
@@ -245,9 +279,11 @@ public class CartEndPoint extends BaseEndpoint {
                 return ResponseService.generateErrorResponse("One or more Serivces not initialized", HttpStatus.INTERNAL_SERVER_ERROR);
             }
             Customer customer = customerService.readCustomerById(customerId);
+
             if (customer == null) {
                 return ResponseService.generateErrorResponse("customer does not exist", HttpStatus.NOT_FOUND);
             }
+                CustomCustomer customCustomer=entityManager.find(CustomCustomer.class,customerId);
             Order cart = orderService.findCartForCustomer(customer);
             if (cart == null)
                 return ResponseService.generateErrorResponse("Cart not found", HttpStatus.NOT_FOUND);
@@ -260,7 +296,7 @@ public class CartEndPoint extends BaseEndpoint {
                     Product product = findProductFromItemAttribute(orderItem);
                     CustomProduct customProduct=entityManager.find(CustomProduct.class,product.getId());
                     if (product != null) {
-                        Map<String, Object> productDetails = sharedUtilityService.createProductResponseMap(product, orderItem);
+                        Map<String, Object> productDetails = sharedUtilityService.createProductResponseMap(product, orderItem,customCustomer);
                         products.add(productDetails);
                         if(customProduct!=null)
                             platformFee=platformFee+customProduct.getPlatformFee();
@@ -343,50 +379,54 @@ public class CartEndPoint extends BaseEndpoint {
 
     @Transactional
     @RequestMapping(value = "place-order/{customerId}", method = RequestMethod.POST)
-    public ResponseEntity<?> placeOrder(@PathVariable long customerId) {
+    public ResponseEntity<?> placeOrder(@PathVariable long customerId,@RequestBody Map<String,Object>map) {
         try {
             Long id = Long.valueOf(customerId);
+            List<Long>orderItemIds=getLongList(map,"orderItemIds");
             if(id==null)
                 return ResponseService.generateErrorResponse("Customer Id not specified",HttpStatus.BAD_REQUEST);
             Map<String, Object> responseMap = new HashMap<>();
             List<Order> individualOrders = new ArrayList<>();
             Customer customer = customerService.readCustomerById(customerId);
-            if (customer == null)
-                return ResponseService.generateErrorResponse("Customer not found", HttpStatus.NOT_FOUND);
             CustomCustomer customCustomer=entityManager.find(CustomCustomer.class,customerId);
+            if (customer == null|| customCustomer==null)
+                return ResponseService.generateErrorResponse("Customer not found", HttpStatus.NOT_FOUND);
             Order cart = orderService.findCartForCustomer(customer);
             if (cart == null)
                 return ResponseService.generateErrorResponse("Cart not found", HttpStatus.NOT_FOUND);
             if(cart.getOrderItems().isEmpty())
                 return ResponseService.generateErrorResponse("Cart is empty",HttpStatus.NOT_FOUND);
             for (OrderItem orderItem : cart.getOrderItems()) {
-                    Product product = findProductFromItemAttribute(orderItem);
-                    Order individualOrder = orderService.createNamedOrderForCustomer(orderItem.getName(), customer);
-                    individualOrder.setCustomer(customer);
-                    if (customer != null)
+                    if(orderItemIds.contains(orderItem.getId())) {
+                        Product product = findProductFromItemAttribute(orderItem);
+                        Order individualOrder = orderService.createNamedOrderForCustomer(orderItem.getName(), customer);
+                        individualOrder.setCustomer(customer);
                         individualOrder.setEmailAddress(customer.getEmailAddress());
-                    individualOrder.setStatus(new OrderStatus("ORDER_PLACED", "order placed"));
-                    OrderItemRequest orderItemRequest = new OrderItemRequest();
-                    orderItemRequest.setProduct(product);
-                    orderItemRequest.setOrder(individualOrder);
-                    orderItemRequest.setQuantity(1);
-                    orderItemRequest.setCategory(product.getCategory());
-                    orderItemRequest.setItemName(product.getName());
-                    Map<String, String> atrtributes = orderItemRequest.getItemAttributes();
-                    atrtributes.put("productId", product.getId().toString());
-                    orderItemRequest.setItemAttributes(atrtributes);
-                    OrderItem orderItemForIndividualOrder = orderItemService.createOrderItem(orderItemRequest);
-                    individualOrder.addOrderItem(orderItemForIndividualOrder);
-                    entityManager.persist(individualOrder);
-                    individualOrders.add(individualOrder);
+                        individualOrder.setStatus(new OrderStatus("ORDER_PLACED", "order placed"));
+                        OrderItemRequest orderItemRequest = new OrderItemRequest();
+                        orderItemRequest.setProduct(product);
+                        orderItemRequest.setOrder(individualOrder);
+                        orderItemRequest.setQuantity(1);
+                        orderItemRequest.setCategory(product.getCategory());
+                        orderItemRequest.setItemName(product.getName());
+                        Map<String, String> atrtributes = orderItemRequest.getItemAttributes();
+                        atrtributes.put("productId", product.getId().toString());
+                        orderItemRequest.setItemAttributes(atrtributes);
+                        OrderItem orderItemForIndividualOrder = orderItemService.createOrderItem(orderItemRequest);
+                        individualOrder.addOrderItem(orderItemForIndividualOrder);
+                        entityManager.persist(individualOrder);
+                        individualOrders.add(individualOrder);
+                    }
                 }
                 responseMap.put("Orders", individualOrders);
                 List<OrderItem> items = cart.getOrderItems();
                 Iterator<OrderItem> iterator = items.iterator();
                 while (iterator.hasNext()) {
                     OrderItem item = iterator.next();
-                    iterator.remove();
-                    entityManager.remove(item);
+                    if(orderItemIds.contains(item.getId())) {
+                        iterator.remove();
+                        entityManager.remove(item);
+                    }
                 }
                 entityManager.merge(cart);
                 return ResponseService.generateSuccessResponse("Order Placed", cart.getId(), HttpStatus.OK);
@@ -413,7 +453,7 @@ public class CartEndPoint extends BaseEndpoint {
                 List<Map<String,Object>>productList=new ArrayList<>();
                 for(Product product:customCustomer.getCartRecoveryLog())
                 {
-                    productList.add(sharedUtilityService.createProductResponseMap(product,null));
+                    productList.add(sharedUtilityService.createProductResponseMap(product,null,customCustomer));
                 }
                 return ResponseService.generateSuccessResponse("Cart Recovery Log : ",productList,HttpStatus.OK);
 
@@ -437,5 +477,7 @@ public class CartEndPoint extends BaseEndpoint {
         System.out.println(product.getName());
         return product;
     }
-
+    public class OrderRequest {
+        private List<Long> orderItemIds;
+    }
 }
